Fix [#133] 채팅링크 연결
